refactoring in hopes of passing some old tests
