+++ lot/landmapper/map_layers/views.py
@@ -2,6 +2,7 @@ from django.conf import settings
 from landmapper import views as lm_views
 from landmapper.models import Taxlot, SoilType, ForestType
 from django.contrib.gis.geos import Point, Polygon, MultiPolygon
+from landmapper.map_layers.utilities import get_bbox_as_string, get_bbox_as_polygon
 
 import geopandas as gpd
 import io, pyproj, shapely, json
@@ -1177,166 +1178,6 @@ def image_result_to_PIL(image_data, alt_size=False):
                 rgba_image = Image.new("RGBA", (settings.REPORT_MAP_WIDTH, settings.REPORT_MAP_HEIGHT), (255,255,255,0))
 
     return rgba_image
-
-def get_bbox_from_property(property, alt_size=False):       # TODO: This should be a method on property instances called 'bbox'
-    # """
-    # PURPOSE:
-    # -   Given a Property instance, return the bounding box and preferred report orientation
-    # IN:
-    # -   property: An instance of landmapper's Property model
-    # OUT:
-    # -   (bbox[str], orientation[str]): A tuple of two strings, the first is the
-    # -       string representation of the bounding box (EPSG:3857), the second is
-    # -       the type of orientation the resulting report should use to print
-    # -       itself
-    # """
-    if hasattr(property, 'geometry_final') and property.geometry_final:
-        geometry = property.geometry_final
-    elif hasattr(property, 'geometry_orig') and property.geometry_orig:
-        geometry = property.geometry_orig
-    elif hasattr(property, 'envelope'):
-        geometry = property
-    else:
-        print("ERROR: property type unrecognized")
-        return ("ERROR", "ERROR")
-    if not geometry.srid == 3857:
-        geometry.transform(3857)
-    envelope = geometry.envelope.coords
-    # assumption: GEOSGeom.evnelope().coords returns (((W,S),(E,S),(E,N),(W,N),(W,S),),)
-    west = envelope[0][0][0]
-    south = envelope[0][0][1]
-    east = envelope[0][2][0]
-    north = envelope[0][2][1]
-
-    property_width = abs(float(east - west))
-    property_height = abs(float(north - south))
-
-    property_ratio = property_width / property_height
-
-    if abs(property_ratio) > 1.0 and settings.REPORT_SUPPORT_ORIENTATION:  # wider than tall and portrait reports allowed
-        if alt_size:
-            target_width = settings.REPORT_MAP_ALT_WIDTH
-            target_height = settings.REPORT_MAP_ALT_HEIGHT
-            orientation = 'portrait'
-        else:
-            target_width = settings.REPORT_MAP_HEIGHT
-            target_height = settings.REPORT_MAP_WIDTH
-            orientation = 'portrait'
-
-    else:
-        if alt_size:
-            target_width = float(settings.REPORT_MAP_ALT_WIDTH)
-            target_height = float(settings.REPORT_MAP_ALT_HEIGHT)
-            orientation = 'landscape'
-        else:
-            target_width = float(settings.REPORT_MAP_WIDTH)
-            target_height = float(settings.REPORT_MAP_HEIGHT)
-            orientation = 'landscape'
-
-    target_ratio = target_width / target_height
-    pixel_width = target_width * (1 - settings.REPORT_MAP_MIN_BUFFER)
-    pixel_height = target_height * (1 - settings.REPORT_MAP_MIN_BUFFER)
-
-    # compare envelope ratio to target image ratio to make sure we constrain the correct axis first
-    if property_ratio >= target_ratio: # limit by width
-        coord_per_pixel = property_width / pixel_width
-        buffer_width_pixel = target_width * settings.REPORT_MAP_MIN_BUFFER / 2
-        buffer_width_coord = abs(buffer_width_pixel * coord_per_pixel)
-        property_pixel_height = property_height / coord_per_pixel
-        height_diff = target_height - property_pixel_height
-        buffer_height_pixel = height_diff / 2
-        buffer_height_coord = abs(buffer_height_pixel * coord_per_pixel)
-    else:   # limit by height
-        coord_per_pixel = property_height / pixel_height
-        buffer_height_pixel = target_height * settings.REPORT_MAP_MIN_BUFFER / 2
-        buffer_height_coord = abs(buffer_height_pixel * coord_per_pixel)
-        property_pixel_width = property_width / coord_per_pixel
-        width_diff = target_width - property_pixel_width
-        buffer_width_pixel = width_diff / 2
-        buffer_width_coord = abs(buffer_width_pixel * coord_per_pixel)
-
-    box_west = west - buffer_width_coord
-    box_east = east + buffer_width_coord
-    box_north = north + buffer_height_coord
-    box_south = south - buffer_height_coord
-
-
-    return (get_bbox_as_string([box_west,box_south,box_east,box_north]), orientation)
-
-def get_bbox_as_string(bbox):
-    # """
-    # PURPOSE:
-    # -   if bbox is not a string of 4 coordinates ("W,S,E,N") convert
-    # IN:
-    # -   A bounding box description, either a string, a list, or a GEOSGeometry
-    # OUT:
-    # -   A string representation of the bbox in the format of "W,S,E,N"
-    # """
-    from django.contrib.gis.geos import Polygon, MultiPolygon
-
-    if type(bbox) in [Polygon, MultiPolygon]:
-        west = bbox.envelope.coords[0][0][0]
-        south = bbox.envelope.coords[0][0][1]
-        east = bbox.envelope.coords[0][2][0]
-        north = bbox.envelope.coords[0][2][1]
-        return "%s,%s,%s,%s" % (west, south, east, north)
-
-    if type(bbox) in [tuple, list]:
-        if len(bbox) == 1 and len(bbox[0]) == 5: # assume coords from Polygon
-            west = bbox[0][0][0]
-            south = bbox[0][0][1]
-            east = bbox[0][2][0]
-            north = bbox[0][2][1]
-            return "%s,%s,%s,%s" % (west, south, east, north)
-        elif len(bbox) == 1 and len(bbox[0]) == 1 and len(bbox[0][0]) == 5:   # coords from MultiPolygon
-            west = bbox[0][0][0][0]
-            south = bbox[0][0][0][1]
-            east = bbox[0][0][2][0]
-            north = bbox[0][0][2][1]
-            return "%s,%s,%s,%s" % (west, south, east, north)
-        elif len(bbox) == 1 and len(bbox[0]) == 1 and len(bbox[0][0]) > 5:  # Non-envelope MultiPolygon coords
-            return get_bbox_as_string(Polygon( bbox[0][0] ).envelope)
-        elif len(bbox) == 1 and len(bbox[0]) > 5:   # Non-envelope Polygon coords
-            return get_bbox_as_string(Polygon(bbox[0]).envelope)
-        elif len(bbox) == 4:    # Simple list [W,S,E,N]
-            return ','.join([str(x) for x in bbox])
-        elif len(bbox) == 2 and len(bbox[0]) == 2 and len(bbox[1]) == 2: # list of two sets of Point coordinates
-            return ','.join([
-                ','.join([str(x) for x in bbox[0]]),
-                ','.join([str(x) for x in bbox[1]])
-            ])
-
-    if type(bbox) == str:
-        if len(bbox.split(',')) == 4:
-            return bbox
-
-    print('ERROR: Format of BBOX unrecognized. Crossing fingers and hoping for the best...')
-    return bbox
-
-def get_bbox_as_polygon(bbox, srid=3857):
-    # """
-    # PURPOSE:
-    # -   return a polygon bounding box from a large number of possible formats
-    # IN:
-    # -   bbox: (str, GEOS geom, list or tuple) any one of the bbox representations
-    # -       of a bounding box supported by 'get_bbox_as_string'
-    # -   ssrid: (int) the EPSG spatial reference identifier for the coordinate
-    # -       system used to interpret the given coordinates
-    # -       default: 3857
-    # OUT:
-    # -   bbox_geom: (GEOS Polygon) A GEOS Polygon feature of the provided bounding
-    # -       box
-    # """
-    from django.contrib.gis.geos import Polygon
-    # Get BBOX into consistant format
-    bbox_str = get_bbox_as_string(bbox)
-    # convert list into W,S,E,N components
-    [west, south, east, north] = [float(x) for x in bbox_str.split(',')]
-    # Format W,S,E,N into point coordinates
-    polygon_input = [(west,south),(east,south),(east,north),(west,north),(west,south)]
-    # Create the Polygon with the provided SRID
-    bbox_geom = Polygon(polygon_input, srid=srid)
-    return bbox_geom
 
 def get_collection_from_objects(source_objects, geom_field, bbox, attrs=[]):
     xmin, ymin, xmax, ymax = bbox.split(',')
